+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Initial implementation of Prometheus exporter.
 - Initial version of Application Insights exporter implemented.
 - Zipkin exporter implemented.
 - Initial version of SDK published. It is based on contribution from Pivotal